Move validate
